requirements.txt
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
langgraph==0.2.28
openai==1.45.0
faiss-cpu==1.8.0
chromadb==0.5.5
sentence-transformers==3.1.0
requests==2.32.3
python-dotenv==1.0.1
tiktoken==0.7.0
pypdf==4.3.1
pandas==2.2.2
mcp==1.0.0
httpx==0.27.2
